requirements_saml.txt
pysaml2>=4.4.0
djangosaml2>=0.16.10
